continuing to consolidate and organize indicators.serializers_new
