+++ indicators/serializers_new/serializers_new_old.py
@@ -49,108 +49,6 @@ class DecimalDisplayField(serializers.DecimalField):
         return None
 
 
-class DisaggregationBase:
-    name = serializers.CharField(source='disaggregation_type')
-    labels = serializers.SerializerMethodField()
-
-    class Meta:
-        model = DisaggregationType
-        fields = [
-            'pk',
-            'name',
-            'labels',
-        ]
-
-    def get_labels(self, disagg):
-        labels = self.context.get('labels_map', {}).get(disagg.pk, None)
-        if labels is None:
-            raise NotImplementedError("no prefetch disaggs")
-        return [{'pk': label.pk, 'name': label.label} for label in labels]
-
-class IPTTDisaggregationMixin:
-    has_results = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'has_results',
-        ]
-
-    def get_has_results(self, disagg):
-        if disagg.pk in self.context.get('with_results', []):
-            return True
-        return False
-
-IPTTDisaggregationSerializer = get_serializer(
-    IPTTDisaggregationMixin,
-    DisaggregationBase,
-)
-
-class IPTTExcelIndicatorMixin:
-    program_pk = serializers.IntegerField(source='program_id')
-    number = serializers.SerializerMethodField(method_name='get_long_number')
-    disaggregations = serializers.SerializerMethodField()
-    no_rf_level = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'program_pk',
-            'number',
-            'is_cumulative_display',
-            'disaggregations',
-            'no_rf_level',
-        ]
-
-    def _get_rf_long_number(self, indicator):
-        level_set = self.context.get('levels', indicator.program.levels.all())
-        level = [l for l in level_set if l.pk == indicator.level_id]
-        if not level:
-            return None
-        level_depth, display_ontology = self._get_level_depth_ontology(level[0], level_set)
-        leveltier = [t for t in self.context.get(
-            'tiers', indicator.program.level_tiers.all()
-            ) if t.tier_depth == level_depth]
-        if not leveltier:
-            leveltier_name = u''
-        else:
-            leveltier_name = u'{} '.format(_(leveltier[0].name))
-        return u"{}{}{}".format(
-            leveltier_name, display_ontology, self._get_level_order_display(indicator)
-        )
-
-    def get_disaggregations(self, indicator):
-        disaggregation_map = self.context.get('disaggregations')
-        disaggregation_labels = {}
-        disaggregation_objects = []
-        disagg_list = self.context.get('disaggregations_indicators').get(indicator.pk, {})
-        for disaggregation_dict in [
-            disaggregation_map.get(d_pk) for d_pk in disagg_list.get('all', [])
-        ]:
-            disaggregation = disaggregation_dict.get('disaggregation', None)
-            labels = disaggregation_dict.get('labels', [])
-            disaggregation_objects.append(disaggregation)
-            disaggregation_labels[disaggregation.pk] = labels
-        disaggregation_context = {
-            **self.context.get('disaggregation_context', {}),
-            'labels_map': disaggregation_labels,
-            'with_results': disagg_list.get('with_results', []),
-        }
-        return sorted(
-            IPTTDisaggregationSerializer(
-                disaggregation_objects, context=disaggregation_context, many=True
-                ).data,
-            key=operator.itemgetter('name')
-        )
-
-
-    def get_no_rf_level(self, indicator):
-        return (not indicator.results_framework or not indicator.level_id)
-            
-
-IPTTExcelIndicatorSerializer = get_serializer(
-    IPTTExcelIndicatorMixin,
-    IndicatorMeasurementMixin,
-    IndicatorBase
-)
 
 class IPTTExcelReportIndicatorBase:
     lop_period = serializers.SerializerMethodField()
@@ -494,154 +392,3 @@ IPTTTPReportIndicatorSerializer = get_serializer(
     IPTTReportIndicatorMixin,
     IndicatorBase
 )
-
-
-class LevelBase:
-    ontology = serializers.SerializerMethodField()
-    tier_name = serializers.SerializerMethodField()
-    name = serializers.SerializerMethodField()
-
-    class Meta:
-        model = Level
-        fields = [
-            'pk',
-            'name',
-            'ontology',
-            'tier_name',
-            'parent_id'
-        ]
-
-    def _get_tiers(self, level):
-        if hasattr(self, 'context') and 'tiers' in self.context:
-            return self.context['tiers']
-        return level.program.level_tiers.all()
-
-    def _get_levels(self, level):
-        if hasattr(self, 'context') and 'levels' in self.context:
-            return self.context['levels']
-        return level.program.levels.all()
-
-    def _get_parent(self, level):
-        if level.parent_id is not None:
-            return [lvl for lvl in self._get_levels(level) if lvl.pk == level.parent_id][0]
-        return None
-
-    def _get_level_depth(self, level):
-        depth = 1
-        target = self._get_parent(level)
-        while target is not None:
-            depth += 1
-            target = self._get_parent(target)
-        return depth
-
-    def _get_level_tier(self, level):
-        tiers = self._get_tiers(level)
-        if len(tiers) > self._get_level_depth(level) - 1:
-            return tiers[self._get_level_depth(level) - 1]
-        return None
-
-    def get_ontology(self, level):
-        target = level
-        ontology = []
-        while self._get_parent(target) is not None:
-            ontology = [str(target.customsort)] + ontology
-            target = self._get_parent(target)
-        return '.'.join(ontology)
-
-    def get_tier_name(self, level):
-        if self._get_level_tier(level):
-            return _(self._get_level_tier(level).name)
-        return None
-
-    def get_name(self, level):
-        return level.name
-
-
-class RFLevelOrderingLevelMixin:
-    indicator_pks = serializers.SerializerMethodField()
-
-    class Meta:
-        override_fields = True
-        fields = [
-            'pk',
-            'indicator_pks'
-        ]
-
-    def get_indicator_pks(self, level):
-        if hasattr(self, 'context') and 'indicators' in self.context:
-            indicators = self.context['indicators']
-        else:
-            indicators = level.program.indicator_set.all()
-        indicators = sorted(
-            [i for i in indicators if i.level_id == level.pk],
-            key=operator.attrgetter('level_order')
-            )
-        return [i.pk for i in indicators]
-
-
-RFLevelOrderingLevelSerializer = get_serializer(RFLevelOrderingLevelMixin, LevelBase)
-
-
-class IPTTLevelMixin:
-    tier_pk = serializers.SerializerMethodField()
-    tier_depth = serializers.SerializerMethodField()
-    chain_pk = serializers.SerializerMethodField()
-
-    class Meta:
-        fields = [
-            'tier_pk',
-            'tier_depth',
-            'chain_pk'
-        ]
-
-    def get_tier_pk(self, level):
-        if self._get_level_tier(level):
-            return self._get_level_tier(level).pk
-        return None
-
-    def get_tier_depth(self, level):
-        if self._get_level_tier(level):
-            return self._get_level_tier(level).tier_depth
-        return None
-
-    def get_chain_pk(self, level):
-        depth = self._get_level_depth(level)
-        target = level
-        if depth == 1:
-            return 'all'
-        while depth > 2:
-            target = self._get_parent(target)
-            depth = self._get_level_depth(target)
-        return target.pk
-
-IPTTLevelSerializer = get_serializer(IPTTLevelMixin, LevelBase)
-
-class IPTTExcelLevelMixin:
-    class Meta:
-        fields = []
-    def get_name(self, level):
-        tier = self.get_tier_name(level) or ""
-        ontology = self.get_ontology(level) or ""
-        ontology = f' {ontology}' if ontology else ""
-        return f'{tier}{ontology}: {level.name}' if tier else level.name
-
-
-IPTTExcelLevelSerializer = get_serializer(IPTTExcelLevelMixin, IPTTLevelMixin, LevelBase)
-
-
-class TierBase:
-    name = serializers.SerializerMethodField()
-
-    class Meta:
-        model = LevelTier
-        fields = [
-            'pk',
-            'name',
-            'tier_depth'
-        ]
-
-    def get_name(self, tier):
-        return _(tier.name)
-
-
-IPTTTierSerializer = get_serializer(TierBase)
